+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version [9.1.0](https://github.com/cedx/free-mobile.js/compare/v9.0.0...v9.1.0)
 - Added the `ClientError` class.
+- Updated the build system to [Gulp](https://gulpjs.com) version 4.
 - Updated the package dependencies.
 
 ## Version [9.0.0](https://github.com/cedx/free-mobile.js/compare/v8.0.0...v9.0.0)